- image db is starting to implement
